+++ appservice/misc.py
@@ -2,10 +2,6 @@ from dataclasses import fields
 from typing import Any
 
 
-class RequestError(Exception):
-    pass
-
-
 def dict_cls(dict_var: dict, cls: Any) -> Any:
     """
     Create a dataclass from a dictionary.
@@ -15,3 +11,19 @@ def dict_cls(dict_var: dict, cls: Any) -> Any:
     filtered_dict = {k: v for k, v in dict_var.items() if k in field_names}
 
     return cls(**filtered_dict)
+
+
+def log_except(fn):
+    """
+    Log unhandled exceptions to a logger instead of `stderr`.
+    Requires the class to have a `logger` variable.
+    """
+
+    def wrapper(self, *args, **kwargs):
+        try:
+            return fn(self, *args, **kwargs)
+        except Exception:
+            self.logger.exception(f"Exception in '{fn.__name__}':")
+            raise
+
+    return wrapper
